+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Removed
-- Remove `Status::setCheckWasCached()` and `Status::wasCheckCached()`. Use the readonly `Status::$wasCached` property instead [#48](https://github.com/orca-services/cakephp-heartbeat/issues/48)
+- **Breaking Change**: Remove `Status::setCheckWasCached()` and `Status::wasCheckCached()`. Use the readonly `Status::$wasCached` property instead [#48](https://github.com/orca-services/cakephp-heartbeat/issues/48)
 
 ### Dependencies