+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Dependencies
 
 ### Fixed
+
+## [4.0.2](https://github.com/orca-services/cakephp-heartbeat/releases/tag/4.0.2) - 2026-08-20
+### Fixed
 - Replace calls to caching methods that no longer exist [#48](https://github.com/orca-services/cakephp-heartbeat/issues/48)
 
 ## [4.0.1](https://github.com/orca-services/cakephp-heartbeat/releases/tag/4.0.1) - 2026-08-20